Fix hit rate calculation to ensure proper type handling of labels

+++ scripts/run_eval.py
@@ -143,7 +143,7 @@ def main() -> None:
             _, _, f1_weighted, _ = precision_recall_fscore_support(
                 labels, preds, average="weighted", zero_division=0
             )
-            hit_rate = float(np.any(preds & labels, axis=1).mean())
+            hit_rate = float(np.any(preds & labels.astype(np.int32), axis=1).mean())
             return {
                 "f1": float(f1),
                 "f1_weighted": float(f1_weighted),
